fix: resolve PHPStan baseline type issues and doc corrections (#50)
* fix: resolve PHPStan baseline type issues and doc corrections

Fix 4 deferred type issues so PHPStan passes clean without baseline
ignores: cast float→int in imagettftext calls, cast int/float→string
in REST header calls, remove dead !empty($values) guard, remove
always-true !$success check. Also fix README listing non-existent
files (notifications, sse), update Requires PHP 7.4→8.3 in plugin
header, rename _fmts→_formats variables, and rename composer script
analyse→analyze.


* docs: correct spelling of "Analyze" in README.md

---------

+++ includes/class-aih-watermark.php
@@ -258,11 +258,11 @@ class AIH_Watermark {
                     for ($x_pos = -$width * 0.5; $x_pos < $width * 1.5; $x_pos += $spacing_x) {
                         // Shadow
                         $shadow = imagecolorallocatealpha($image, 0, 0, 0, 60);
-                        @imagettftext($image, $font_size, 30, $x_pos + 3, $y_pos + 3, $shadow, $font_file, $text);
-                        
+                        @imagettftext($image, $font_size, 30, (int) ($x_pos + 3), (int) ($y_pos + 3), $shadow, $font_file, $text);
+
                         // White text
                         $white = imagecolorallocatealpha($image, 255, 255, 255, 50);
-                        @imagettftext($image, $font_size, 30, $x_pos, $y_pos, $white, $font_file, $text);
+                        @imagettftext($image, $font_size, 30, (int) $x_pos, (int) $y_pos, $white, $font_file, $text);
                     }
                 }
